Record reference transaction ordering in Transaction
Transaction is currently recording all reference updates in a single
map, and verifying all reference updates in a single reference
transaction. This is fine for most operations but some RPC handlers,
mainly fetches, may stage updates that would lead to a directory-file
conflict by for example staging:

1. delete 'refs/heads/parent'
2. create 'refs/heads/parent/child'

To accommodate for such transactions, this commit changes the Transaction
to record the individual reference transactions made. This retains the
ordering knowledge between the reference updates performed and allows for
saying which of the conflicting reference should exist at the end of the
transaction.

Most of the logic around references stays the same. The notable differences
are in reference verification and logging:

1. References are now verified by applying the reference transactions against
   a snapshot of the target repository. This ensures the verification process
   takes into account the correct ordering of the reference transactions, and
   that the directory-file inducing 'refs/heads/parent' is deleted before the
   child reference is verified.
2. The log entries now record reference transactions, and they are applied in
   order to the repository when the log entry is being applied.

With the above changes, it's now possible to perform updates in a transaction
which would have before resulted a directory-file conflict.