Refactor Snapshot as SnapshotLSN
The Snapshot type contained previously more fields. With the introduction
of the file system level snapshots, we no longer had to track for example
the version of the hooks that should be used in the transaction. Nowadays
the Snapshot only contains the read index field, which is the log index the
transaction is reading the partition's state at. Simplify by removing the
Snapshot type and replacing it with snapshotLSN field that directly records
the LSN the transaction is reading at.