Decouple snapshotting logic from Transaction
The snapshotting logic is currently tied to the Transaction. This was
fine until now as we always created a single snapshot in context of a
transaction. We're soon about to verify transactions against a
snapshot of the target repository. In order to reuse the code to do so,
this commit extracts out the snapshotting logic into a separate type.
The logic remains the same but we no longer access fields directly
from TransactionManager nor Transaction, and the snapshot type helps
with rewriting the relative paths to point to the snapshot.