Don't error if deleting test repository fails
Our CreateRepository helper has clean up logic in place to remove
a repository automatically at the end of a test. Currently it fails
the test if deleting the repository errors. This is undesired as some
tests leave the repository in invalid states where it can't be deleted
any more. Instead of failing the test, this commit changes the helper
to log the error. It's debatable whether it's even necessary to remove
the repository at the end of the test given the helper creates the
repository in the tests storage directory which is removed at the end
of the test. Do the smaller change now anyway by just logging the
failure instead.