helper/perm: Drop log message when `os.FileInfo` is `nil` on walk
It's perfectly legit for the `os.FileInfo` that is passed to the
callback function in `filepath.Walk()` to be `nil`. The most common
scenario where this happens is when the file has been concurrently
removed, which is a benign edge case.

Let's drop the error message. It doesn't add any meaningful information
and the logged errors would be benign anyway.