proto: Add linter rules for comments
Currently we don't have any linting rules for comments in proto files.
This has led to various styled being used in the comments and lack of
consistency. Enforce the styling for comments in proto files.

This commit also includes manual changes to comments in proto files to
adhere to these new rules. The changes are lazy in nature, mostly
minimal changes to simply conform to the new rules. This is to reduce
review load and also writing good comments is out of the purpose of this
commit.