Fix flaky PartitionManager test
TestPartitionManager/starting_transaction_failed_due_to_cancelled_context
is asserting that context cancellation is taken into account while beginning
a transaction. The test is currently flaky as it may be that the
TransactionManager finishes initializing before Begin() is called. This then
triggers indeterministic behavior in Begin(). Either the transaction begins
successfully, or the context cancelation is noticed and it errors out. Fix
this flake by preventing the TransactionManager from running in the test so
it never initializes and the Begin will time out.