Add a method for getting the original repository of the request
Some functionality in Gitaly would benefit from having the original
repository the request contained before rewriting it to point to the
transaction's snapshot. This is mostly locations that store some state
that is shared across requests. Generally these locations would be
caches and they use the relative path of the repository as a cache
key. The cache key should remain stable even if the repository's
relative path is rewritten. This commit adds OriginalRepository
method on the transaction to acquire the repository as it came in the
request. This can then be used to construct stable cache keys in
various locations.

While this allows for keeping the caches working as is, we should
separately evaluate how the caches interact with transactions.