Use original relative path when building cache keys
Transactions rewrite the repository in the request to point to the
transaction's snapshot of the repository. This means the relative
path in the request is not stable between different requests which
would lead to Gitaly's caches always missing. This commit extracts
the original relative path from the request and uses it to build the
cache keys in pack-objects and info-refs caches. This way the cache
keys remains stable between different transactions.