Setup repository correctly in TestCoordinator_grpcErrorHandling
TestCoordinator_grpcErrorHandling is not creating the test repository
on the Gitaly storages configured as part of a virtual storage. This
causes the test to fail when WAL is enabled as the UserCreateBranch
errors with unexpected error due to the repository not existing. Setup
the repository in the storages to avoid this.