protogem: Add URLs to the homepage, source code and bug tracker
Even though we don't link to the location where the source code of our
Ruby Gem exists, rubygems.org does have that information set. It points
to the legacy location of our Ruby Gem though, which is hosted in the
gitlab-org/gitaly-proto repository. It's been ages since we have
deprecated that repository though, so it's kind of puzzling that the
site still has that information. My best guess is that it simply uses
information that was at one point in time present in our Gemspec.

Let's explicitly document these URIs as part of our metadata. As
documented in the specification [1], these URIs should indeed end up on
rubygems.org and thus fix the out-of-date links.

[1]: https://guides.rubygems.org/specification-reference/#metadata