Take in storage and relative path instead of Repository in PartitionManager
PartitionManager is currently taking in a storage.Repository to determine
the target repository of a transaction. This worked reasonably for a single
repository. It's not as convenient as an interface when a transaction may
span multiple repositories. Each of the repositories must be in the same
storage, so storage should be given as a parameter only once. The quarantine
parameters may also be set in Repository yet they are completely ignored
by the transaction code. Change the interface to instead take in a storage
name and the relative path of the repository directly. This removes the
unnecessary fields inside Repository from the interface. When we start
supporting multi-repo transactions, we'll only take in the storage name
once and can take in the other relative paths a simple slice.