Standardize on referring to log positions with LSN
LSN or log sequence number is the usual term used to refer to a given
position in a log. Our code is currently using log index in most places.
This commit standardizes the name to LSN by renaming each use of log index
to LSN. This way we'll have easier time also writing documentation later
that refers to LSNs.