Test WAL with SHA256 in CI
We're currently running TransactionManager's tests with SHA256 as
part of the SHA256 enabled tests. We're not running the `test-wal`
and `test-with-praefect-wal` jobs with SHA256. Those jobs will wire
up transaction management into the test server and run all of our
service tests with WAL enabled. Let's run them also as part of the
SHA256 tests to ensure the service tests work fine with WAL and
SHA256.