praefect: Get repository ID of repos in database
Currently the `track-repositories` Praefect command checks if the target
repo is already tracked in the database, but does not retrieve the
repository ID in this scenario, returning 0 instead. As a result,
track requests will fail when we use the invalid repository ID returned.

Fetch the actual repository ID in this scenario.

Changelog: fixed