backup: Support incremental server-side backups
Incremental backups are currently supported only for client-side
repository backups, i.e. if the `REPOSITORIES_SERVER_SIDE=true` env
var is not set when running the gitlab:backup:create Rake task. This is
because we're not providing this flag in the request to the
BackupRepository RPC. This commit:

- Modifies the BackupRepositoryRequest message to include an
  `incremental` flag and update the surrounding documentation.
- Threads the `incremental` flag through the gRPC request and into the
  backup manager on the other end.
- Updates some of the tests to exercise the flag.

Changelog: added