backup: Set default branch from manifest on restore
Adds a test for the worst case scenario where the backup was of an empty
repository. In this case there is no bundle file to derive the default
branch.

Eventually manifest files will have blank paths to indicate, for
example, that there is no bundle or no hooks. Until then the code
expects that all paths will be set, even if the destination file does
not exist.

Changelog: changed