Fix TestUserCreateBranch_successful writing into shared repository
TestUserCreateBranch_successful is setting up a shared repository in
the tests. It then proceeds to concurrently write to the shared
repository in the subtests. This leads to indeterministic results as
the branch may or may not be already created by one of the other
subtests. Fix this by running each of the tests agains their own
repository. As all tests are now running agains their own repositories,
we know longer need to delete the created branch at the end of the test.