git: Bump minimum version of Git to 2.42
Let's bump the minimum version of Git to 2.42 with this bring in
the following important changes:

* git-pack-refs(1) learned to take --include and --exclude, which allows
us to skip packing certain references. This is important in the context
of the write-ahead  log so that we don't pack write-ahead log
references.
* SHA256 is no longer marked as experimental.
* Git has learned to more efficiently exclude references, which should
help repositories which have many hidden references.
* git-rev-list(1) has learned to take pseudo-refs like --all and --not
in --stdin mode.
* git-cat-file(1) has learned a new -Z switch, which uses NUL delimiters
for both stdin and stdout.