cgroups: Create repository cgroups on-demand
On systems with very large numbers of existing cgroups, initializing
Gitaly's repository cgroups may take upwards of 20 seconds. With
`tableflip` this delay has no impact as the existing process is still
present to take traffic until the new one finishes bootstrapping.
However, we need to remove `tableflip` as the WAL does not allow
concurrent Gitaly processes accessing the same repository.

Create a given repository cgroup on-demand the first time a
process needs to be added to it.

Changelog: performance