Make relative path a mandatory test parameter
TransactionManager tests are currently defaulting to the pre-existing
repository's relative path if one is not specified in the test. This
takes away the empty relative path as a valid parameter in the test
setup. Configure each test case explicitly with the relative path they
want to access instead. This ensures:

1. We can explicitly test the case where transaction is begun with
   an empty relative path.
2. Lays ground for future refactorings to remove the pre-existing
   repository from the tests. It was needed before as we couldn't
   create repositories through transactions. Now we can, so we should
   exercise that code where we need a repository. This is left for
   the future though.