Implement general ticker factory helpers
Gitaly's code is using tickers in multiple places. They make it easier
to test code that runs on an interval. The tickers are generally
constructed using a factory method injected into the type using the
ticker. This makes it easier to mock out the timer based tickers in
tests. We'll soon need this in the PartitionManager tests. This commit
implements factory methods for timer ticker and a ticker that never
ticks. The timer tickers can be used to configure the production code,
and the null tickers can be used to disable the ticking for example in
tests. If the tests need to manually control the ticking, they can
do so by passing a custom TickerFactoryFunc into the type under test.