Only snapshot valid Git directories
Transaction manager currently includes all requested relative paths
in the snapshot. Typically this would be the relative path of the
target repository. If the client fed it paths such as '@hashed', it
would attempt snapshot the entire storage. If Gitaly controlled the
directory structure, we could easily guard against this by only
allowing leaf/project directories to be snapshotted. The paths are
currently controlled by the client so we can't rely on this without
possibly breaking some paths.

Workaround this by checking whether the snapshotted directory is a
git repository. If it isn't, the snapshotting will error out. This
prevents using any of the higher level directories as the path to
snapshot.

The error equality assertion when beginning a transaction was changed
to a an error.Is assertion. This avoids having to repeat the error
annotations in tests which are not meaningful while still testing that
the expected error is in the chain.