Support partitioning hints
ObjectPool related RPCs often carry in the request the target repository
and the alterante repository. Depending on the RPC the roles switch but
these refer to the object pool and the member repository.

Gitaly is currently assigning relative paths into partitions on first
access. As it is, each repository goes into its own partition unless
it has an alternate. If there is an alternate, the repository is assigned
to the same partition. This ensures the pool and the member are assigned
into the same partition so their operations are serialized.

We currently don't support assigning a repository explicitly into the
same partition as another repository. This will be needed in order to
ensure that a repository that is being linked to a pool will be assigned
to the same partition as the pool even before the link is made. Similarly,
this can be used to ensure the repository created with CreateFork will be
assigned to the same partition as the pool.