Include alternate repo in the snapshot
Some object pool related RPCs carry an alternate repository in the
request they access. This repository needs to be included in the
transaction's snapshot in order to ensure the RPC can access it as
expected. This commit extracts the alternate repository from the
request and includes it also in the transaction.