Enable transactions for FetchIntoObjectPool
This commit enables transactions for FetchIntoObjectPool. Some tests
were updated as failure scenarios may before the RPC handler is invoked
and thus have different errors.

The RPC handler is pruning broken references. The tests are updated to
not assert these are removed. Gitaly is not voting on these changes with
Praefect, and is not committing them via the WAL. Committing them via the
WAL wouldn't work either as updating such references would fail. Such
references would have to be manually cleaned up.