Enable transactions for CreateObjectPool
This commit enables transactions for CreateObjectPool. One test
had to be changed as it now fails earlier in the call chain due to
invalid state and leads to a different error messsage. Another test
asserting lock semantics is skipped. RPCs won't see each other's
locks with transactions.

TestCreate is now getting the test commit created in a branch where
as it was previously unreferenced. Object pool creations are treated
as normal repository creations by transactions. As such, unreferenced
objects are not included in the write-ahead logged packs. Referencing
the object allows the later assertions for object existence to pass.
Overall, this shouldn't really make a big difference in behavior. If
nothing is referencing the objects anyway, the pool doesn't need them.
If they are later referenced, later fetches to the pool will retrieve
them. The bigger question is whether packing a repository in the single
pack that gets logged deteroriate performance meaningfully. If this
becomes a problem, we can later special case object pool creations and
write-ahead log them more efficiently.