Enable transactions for LinkRepositoryToObjectPool
This commit enables transactions for LinkRepositoryToObjectPool.
This requires just recording the updated alternate link in the
transaction.

Transactions require the object pool and the member repository to
be in the same partition. This may not be the case when Praefect
attempts to reconciliate alternate link differences during
replication. Moving repositories between partitions is difficult, so
for now we'll simply ignore the pool linking failure in Praefect and
leave accept deviating alternate links.