limiter: Attach some stats in the reported backoff events
When the resource watchers report the backoff events to the adaptive
calculator, they attach a reason for the backoff (threshold exceeded for
example). The reason consists of some stringified numbers for
clarification. This commit moves those numbers to a hashmap for better
readability and explorability. In some later commits, we'll attach more
stats to that hashmap.