Set logging level config to "info" explicitly
In multiple places, we state that the default log level is "info" by
default. The "info" severity is set when the underlying logging library
parses the log level unsuccessfully. We don't have set the default
config to "info".

In a prior commit, we fixed a grpc logger issue due to the missing of
configured log level. To prevent that from happening again at another
place, this commit sets a default logging level.