cgroups: Unify TestStats
Combine the v1 and v2 versions of `TestStats()` to a single test.

Here we have the opposite of `TestPruneOldCgroups()`, the test cases
are unique to v1 or v2, but the assertions are version agnostic. Add a
`version` field to the test cases, rather than iterating over each case
with each version as we did with the other tests.