limiter: Make CpuThrottledThreshold configureable
In the current implementation, the adaptive limiting kicks in when the
resource level exceeds some hard-coded thresholds:
* 90% of the parent cgroup's memory.
* Cgroup's cpu is throttled for 50% of the observation time.

Although the current CPU throttled threshold is reasonable, it might not
be good for all cases. A more powerful machine can tolerate a higher
throttling rate while a less powerful machine wants to lower the limit
sooner. This commit adds the ability to customize the CPU throttled
threshold.