repository: Remove RenameRepository RPC
Remove the RenameRepository RPC as its no longer being used. With this
we delete all associated code and tests around this RPC and also delete
the proto definitions. While it might have been easier to review as
isolated commits, the code is highly intertwined making isolation
harder, as such the whole deletion is taken care of in this commit.