gitaly: Move asynchronous tasks to end of startup
We have several background tasks spawned as part Gitaly's startup
process. Currently we start many of these towards the beginning of
startup, but on systems with limited available parallelism these may
slow the progression of the main goroutine.

Move these to the end of startup so that critical tasks can complete as
quickly as possible.

On a 2-core host this improves startup time by roughly 15%:

  Benchmark 1: ./async-prune serve config.toml
    Time (mean ± σ):     535.3 ms ± 106.7 ms    [User: 462.0 ms, System: 228.5 ms]
    Range (min … max):   436.9 ms … 727.9 ms    10 runs

  Benchmark 2: ./tasks-end serve config.toml
    Time (mean ± σ):     468.0 ms ±  23.5 ms    [User: 438.6 ms, System: 220.6 ms]
    Range (min … max):   430.8 ms … 517.3 ms    10 runs

  Summary
    ./tasks-end serve config.toml ran
      1.14 ± 0.24 times faster than ./async-prune serve config.toml

Changelog: performance