backup: Track bytes written by LazyWriter
Modifies the LazyWriter so the number of bytes written to the underlying
WriteCloser is tracked. Adds a new public method to retrieve this value
and updates existing tests to exercise this function.

The number of bytes written will later be used to observe the bundle
size metric added in the previous commit.