backup: Rename ManifestLocator to ManifestInteroptLocator
Soon we will introduce a real ManifestLocator that will have no
fallback. The goal is to add manifests for all existing layouts to
smooth this transition.

We can't just modify this existing layout because we want to be able to
add a new locator to the locator option on the gitaly-backup
commandline. This is so that we can coordinate changing tests on
gitlab-rails with the release of the full manifest locator.