log: Add thread safety to logger interface
The `logrus.entry` is not thread safe, so our logger
which uses `logrus.entry` directly is not thread
safe either. And it is breaking the race-go test.

In this commit, we use a wrapper with lock to wrap `logrus.entry`.
The implementation of logger interface is also changed
to be thread safe.