Add Housekeeping to the transaction log entry
This commit adds Housekeeping to the transaction log entry protobuf. Up
to this point, the pack-refs housekeeping task is the only supported
task. The task entry includes the list of pruned references. The
packed-refs file will be copied from the transaction's staging directory
over log entry and then the destination repository. That file's name is
static. Hence, we don't need to store its content at the moment.

When the log entry is applied, Gitaly replaces the existing packed-refs
file by the new one in the log entry. It also clears mentioned loose
references in the disk.