bundleuri: Log if bundle_uri is being used
We want to know if bundle-uri is being used and has been sent along in
the packfile response. This can be useful in RPC logs for
PostUploadPackWithSidechannel and SSHUploadPack, for instance, to tell
whether or not less CPU was used as a result of a client's prefetch of a
bundle through the bundle uri.