errors: Use `errors.As()` instead of switch type assertions
In few places of our code base, we have code like:

switch terr := err.(type) {
case nil:
if err == nil {
	return vs
case ValidationErrors:
	for _, err := range terr {
}

This doesn't work with wrapped errors. So let's use `errors.As()`
instead. This makes the code a little messier but is compatible with
wrapped errors.

We skip making changes in structerr package, since there we explicitly
want to unwrap the layers to extract the metadata.