cleanup: Check parents of internal refs
Currently we check if a commit pointed to by of our internal references
is directly listed on the user-provided commit-map file. However, in
some cases we create internal commits that may have a rewritten commit
as a parent. `refs/merge-requests/<MR_IID>/merge` refs are merge
commits created internally by GitLab to check if an MR can be cleanly
merged and cannot be rewritten by clients. We also create a
`keep-around` ref for each iteration of the `merge` ref that is not
advertised to clients.

These commits will not be present on the commit-map, but they may be
children of a rewritten commit. Without removing the refs for these
isolated merge commits we will fail to remove all of the internal
references that are keeping a large blob alive.

These refs are generally safe to remove, rails does so two weeks after
closing/merging an MR. Should a `merge` ref be removed immediately after
creation and before the mergeability background job can start, this
would require the user to either add an additional commit to the MR,
recreating the ref, or create a new MR.

To ensure we remove all references that connect to rewritten commits,
update our git-for-each-ref(1) format string to include both the object
referenced by the reference and its parents. If any of these are present
on the commit-map then we delete the reference.

Changelog: fixed