proto: Define `CheckBlobsGenerated` RPC
In certain scenarios it is useful for clients to have insight regarding
whether a particular file in a repository is generated. For example, in
the context of rendering diffs, generated files may not be desired. To
accommodate this, a new `CheckBlobsGenerated` RPC is defined. A new
`AnalysisService` gRPC service is also defined which contains the RPC.